move Worker to its own module
